fix error handler

+++ src/logger.js
@@ -1,9 +1,7 @@
 const fs = require("fs");
 const storageFilePath = "./history.log";
 
-if (!fs.existsSync(storageFilePath)) {
-  fs.writeFileSync(storageFilePath);
-}
+if (!fs.existsSync(storageFilePath)) fs.writeFileSync(storageFilePath);
 
 module.exports.addPath = (path) => {
   fs.appendFileSync(storageFilePath, path + "\n");
